clean up triggercheckpoint
clean up triggercheckpoint
