+++ Assets/Scripts/Player/OnCheckpointHit.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ab4cba2b253b47e46888edda23d2c242
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
