--- Assets/Scripts/Player/TriggerCheckpoint.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6deb514ad86ec3d49a3b66d0c7c80cf8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
